+++ src/common/hooks/useIntegrationConfigMutation/useIntegrationConfigMutation.js
@@ -3,10 +3,14 @@ import {
 } from 'react-query';
 import { cloneDeep } from 'lodash';
 
-import { useOkapiKy } from '@folio/stripes/core';
+import { useOkapiKy, useStripes } from '@folio/stripes/core';
+
+import { SCHEDULE_PERIODS } from '../../../OrganizationIntegration/constants';
+import { getUTCDate } from '../../../OrganizationIntegration/utils';
 
 export const useIntegrationConfigMutation = (options = {}) => {
   const ky = useOkapiKy();
+  const stripes = useStripes();
 
   const { mutateAsync } = useMutation({
     mutationFn: (integrationConfig) => {
@@ -20,6 +24,18 @@ export const useIntegrationConfigMutation = (options = {}) => {
       if (scheduleParameters?.weekDays) {
         scheduleParameters.weekDays = Object.keys(scheduleParameters.weekDays)
           .filter(weekDay => scheduleParameters.weekDays[weekDay]);
+      }
+
+      if (scheduleParameters?.schedulePeriod === SCHEDULE_PERIODS.days && scheduleParameters?.schedulingDate) {
+        scheduleParameters.schedulingDate = getUTCDate({
+          date: scheduleParameters.schedulingDate,
+          time: scheduleParameters.scheduleTime,
+          timezone: stripes.timezone,
+        });
+      }
+
+      if (scheduleParameters?.scheduleTime) {
+        scheduleParameters.scheduleTime = scheduleParameters.scheduleTime.slice(0, 8);
       }
 
       const kyMethod = options.method ?? (integrationConfig.id ? 'put' : 'post');